refactor(types): extract common tests to sibling file
